pyproject.toml
[project]
name = "restext"
version = "0.1.0"
description = "Context engine for AI agents"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",

    # Vector store
    "qdrant-client>=1.12",

    # Embeddings
    "openai>=1.50",
    "tiktoken>=0.8",

    # Crawling & parsing
    "httpx>=0.27",
    "trafilatura>=1.12",
    "pymupdf>=1.24",
    "python-docx>=1.1",

    # Background jobs
    "arq>=0.26",
    "redis>=5.0",

    # Auth & validation
    "pydantic>=2.9",
    "pydantic-settings>=2.6",

    # File uploads
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]
